I made a blank of the HomeDummyTable page and registered a route for it.   Test for   Backendless
I made a blank of the HomeDummyTable page and registered a route for it. To do this, in the HomeDummyTable.jsx file, in the HomeDummyTable function, I returned some markup for the HomeDummyTable page blank and exported this HomeDummyTable function. Then, in the index.js file, I re-exported the Home page. Then, in the App.jsx file, I imported Routes, Route and lazy. I also imported the HomeDummyTable component with help the lazy function  . In the App function, I returned <Routes> in which I specified the Route where I specified the <HomeDummyTable />     page.   Test for   Backendless