Added navigation buttons dummyList  ,  dummyTable and   dummyChart   and then  slightly implemented the logic of their operation.    Test for   Backendless
Added navigation buttons dummyList  ,  dummyTable and   dummyChart   and then  slightly implemented the logic of their operation. To do this, I wrote a blank for the  HomeDummyList   component in the   HomeDummyList.jsx   file.  Then, in the index.js file (which is located in the  homeDummyList  folder), I re-exported the   HomeDummyList  component.  I did similar actions for the DummyChart component in the DummyChart.jsx file. And also re-exported it in the index.js file (which is located in the dummyChart folder).  Then, in the Navigation.jsx file, I imported NavLink and, using it in the Navigation function, made transitions along routes when clicked to the   dummyList    ,  dummyTable   or   dummyChart   buttons   .  And also re-exported the Navigation component in the index.js file (which is located in the navigation folder).   Then, in the Layout.jsx file, I imported Suspense and Outlet to use them to render different page bodies in the Layout component.    I also rendered the component
 Navigation in it.   And also re-exported the Layout component in the index.js file (which is located in the layout folder).   In the App.jsx file I commented out the   HomeDummyList    component for now (I will uncomment it and render it in the next commits) I also imported the Layout component, and using lazy I imported the  DummyTable   and   DummyChart    component and   assigned routes to them in Route   .  I also removed the HomeDummyTable component and made the HomeDummyList component for the home page  .   Test for   Backendless   .