I uncommented the import of the HomeDummyList page in the App.jsx file and specified the path for it.   Test for Backendless
I uncommented the import of the HomeDummyList page in the App.jsx file and specified the path for it.   At the same time, I made it so that by default when the page is loaded, the HomeDummyList page opens and at the same time, so that when moving through the pages, the name of the page is added to the route. For example http://localhost:3000/test-backendless/dummyTable or http://localhost:3000/test-backendless/dummyChart   . To do this,   I added dummyList path in Navigation.jsx file   .  Then in the App.jsx file for the HomeDummyList page I set the path to index so that it loads by default when this application is opened. And I also set the dummyList path for the HomeDummyList page so that the path http://localhost:3000/test-backendless/dummyList is displayed in the browser when going to the HomeDummyList page (according to the Technical Specifications).  Also I adjusted the styles in the Navigation.styled.js file.   Test for Backendless