I'm refactoring the code in the App.jsx file  (I'm doing everything I did in previous commits), adjusting it to the technical specifications. Test for Backendless
I'm refactoring the code in the App.jsx file  (I'm doing everything I did in previous commits), adjusting it to the technical specifications. The only difference from the previous code is that I take the paths to the site page files from the tabs.json file   .  To implement this logic, I imported Navigate, useEffect, useState.  Then in useState I set the initial value of the array into which I will subsequently load the data from the tabs.json file.  Then in useEffect I got the data from the tabs.json file and immediately wrote it to the tabsData variable  .  Then in the sortedData variable I sorted the data from the tabs.json file by order value. Then I used map to access the sortedData array data (these are the paths to the site pages that I load using lazy , id et cetera). Also here I applied the condition, if i === 0 (zero index of the array), then this is the home page of the Dummy List. Then using Navigate I will redirect directly to the Dummy List page  .   Test for Backendless