I'm refactoring the code  (I'm doing everything I did in previous commits), adjusting it to the technical specifications. Test for Backendless
I'm refactoring the code  (I'm doing everything I did in previous commits), adjusting it to the technical specifications. To implement this logic, I passed the array obtained from the tabs.json file using props. First it was tabs={sortedData} in the App.jsx file. Then in the Layout.jsx file I accepted this prop and passed it on using dataTabs={tabs} . Then, in the Navigation.jsx file, I accepted this prop and applied map to it to make the necessary markup and insert the values from the array into this markup (this array was obtained from the tabs.json file). Test for Backendless